requirements/test.txt
-r _base.txt

django-debug-toolbar==4.2.0
django-silk==5.0.4
